{"schema":"decodex/commit/1","summary":"Split stale active reentry helpers","authority":"XY-1115"}
